refactor(normalizer): move review.py into scripts/
The review-HTML builder is only used by the dry-run script, so it belongs
alongside it under scripts/ rather than in the normalizer package.
normalize_dryrun.py now imports it as a sibling script. Its unit tests are
dropped (per request) since scripts/ has no parallel test tree; the eval
scorer tests it shared a file with are kept in tests/normalizer/test_evals.py.
